+++ buildbot.py
@@ -56,12 +56,6 @@ def build(properties):
     run_command(command)
 
 
-def docs(properties):
-    command = [sys.executable, 'waf', 'docs',
-               '--all_docs', '--publish_docs', '-v']
-    run_command(command)
-
-
 def run_tests(properties):
     command = [sys.executable, 'waf', '-v', '--run_tests']
     run_cmd = None
@@ -110,8 +104,6 @@ def main():
         run_tests(properties)
     elif cmd == 'install':
         install(properties)
-    elif cmd == 'docs':
-        docs(properties)
     else:
         print("Unknown command: {}".format(cmd))