+++ NEWS.rst
@@ -6,7 +6,8 @@ every change, see the Git log.
 
 Latest
 ------
-* tbd
+* Major: Updating how docs are built
+* Update waf build script
 
 6.3.0
 -----